Marshal encryption and KDF params using protobufs
Adds two additional protobuf messages:
 - `RawNebulaEncryptedData`, a wrapper containing `Ciphertext` and
    `EncryptedMetadata`
 - `RawNebulaEncryptedDataMetadata`, which contains strings denoting
   the `EncryptionAlgorithm` and the `KeyDerivationParameters`

The commit uses the existing encoding format for Argon2id parameters, as
I believe this format should be fairly flexible for handling other KDFs
as well. The main improvement in this diff is that it offers more
flexibility in modifying the structure of metadata in the future, as
well as storing the encryption algorithm used.